pyproject.toml
[project]
name = "competition-bakdabak-action-datamining"
version = "0.1.0"
requires-python = "==3.10.*"
dependencies = [
    "ipykernel>=7.1.0",
    "kaggle>=1.7.4.5",
    "nbconvert>=7.16.6",
    "python-dotenv>=1.0.0",
    "roboflow>=1.2.11",
    "torch>=2.0.0",
    "ultralytics>=8.3.221",
]
